requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
mongoengine==0.24.0
protobuf==3.19.4
pymongo==4.0
six==1.16.0
